+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+* Remove Clndr dependency
+* Remove Moment dependency
+
 ### 1.0.0: July 21th, 2020
 * Plugin boilerplate