+++ CHANGELOG.md
@@ -1,8 +1,9 @@
+### 1.0.1: July 22nd, 2020
 * Remove Clndr dependency
 * Remove Moment dependency
 * Admin settings for each job filter
 * Frontend filters
 * Frontend loading effect
 
-### 1.0.0: July 21th, 2020
+### 1.0.0: July 21st, 2020
 * Plugin boilerplate